Restructured project files into packages for best android dev practice and making sure gitignore behaves properly
