+++ .gitignore
@@ -13,3 +13,6 @@
 .externalNativeBuild
 .cxx
 local.properties
+
+# Google Services (e.g. APIs or Firebase)
+google-services.json
