+++ app/src/main/java/com/example/munch_cmpt362/MunchApplication.kt
@@ -1,0 +1,7 @@
+package com.example.munch_cmpt362
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MunchApplication : Application()
