+++ app/src/main/java/com/example/munch_cmpt362/data/repository/AuthRepository.kt
@@ -1,0 +1,35 @@
+package com.example.munch_cmpt362.data.repository
+
+import com.google.firebase.auth.FirebaseAuth
+import kotlinx.coroutines.tasks.await
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class AuthRepository @Inject constructor(
+    private val auth: FirebaseAuth
+) {
+    val currentUser = auth.currentUser
+
+    suspend fun login(email: String, password: String): Result<Unit> = try {
+        auth.signInWithEmailAndPassword(email, password).await()
+        Result.success(Unit)
+    } catch (e: Exception) {
+        Result.failure(e)
+    }
+
+    suspend fun register(email: String, password: String): Result<Unit> = try {
+        auth.createUserWithEmailAndPassword(email, password).await()
+        Result.success(Unit)
+    } catch (e: Exception) {
+        Result.failure(e)
+    }
+
+    fun logout() {
+        auth.signOut()
+    }
+
+    fun isUserLoggedIn(): Boolean {
+        return auth.currentUser != null
+    }
+}
