+++ app/src/main/java/com/example/munch_cmpt362/di/RepositoryModule.kt
@@ -1,0 +1,20 @@
+package com.example.munch_cmpt362.di
+
+import com.example.munch_cmpt362.data.repository.AuthRepository
+import com.google.firebase.auth.FirebaseAuth
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object RepositoryModule {
+
+    @Provides
+    @Singleton
+    fun provideAuthRepository(auth: FirebaseAuth): AuthRepository {
+        return AuthRepository(auth)
+    }
+}
