+++ app/src/main/java/com/example/munch_cmpt362/ui/MainActivity.kt
@@ -1,4 +1,101 @@
 package com.example.munch_cmpt362.ui
 
-class MainActivity {
+import android.Manifest
+import android.content.pm.PackageManager
+import android.location.Location
+import android.location.LocationListener
+import android.location.LocationManager
+import android.os.Build
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.widget.ViewPager2
+import com.example.munch_cmpt362.R
+import com.example.munch_cmpt362.SwipeFragment.SwipeFragment
+import com.example.munch_cmpt362.ui.adapter.MyFragmentStateAdapter
+import com.google.android.material.tabs.TabLayout
+import com.google.android.material.tabs.TabLayoutMediator
+
+class MainActivity : AppCompatActivity(), LocationListener {
+    private lateinit var swipeFragment: SwipeFragment
+    private lateinit var viewPager: ViewPager2
+    private lateinit var tabLayout: TabLayout
+    private lateinit var myMyFragmentStateAdapter: MyFragmentStateAdapter
+    private lateinit var locationManager: LocationManager
+    private val PERMISSION_REQUEST_CODE = 0
+    private val tabTitles = arrayOf("Swipe", "temp1", "temp2")
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        viewPager = findViewById(R.id.viewpager)
+        viewPager.setUserInputEnabled(false);
+        tabLayout = findViewById(R.id.tab)
+
+        swipeFragment = SwipeFragment()
+        val fragments = arrayListOf<Fragment>(swipeFragment)
+
+        myMyFragmentStateAdapter = MyFragmentStateAdapter(this, fragments)
+        viewPager.adapter = myMyFragmentStateAdapter
+
+        TabLayoutMediator(tabLayout, viewPager) { tab, position ->
+            tab.text = tabTitles[position]
+        }.attach()
+
+        checkPermission()
+    }
+
+    private fun checkPermission() {
+        if (Build.VERSION.SDK_INT < 23) {
+            initLocationManager()
+            return
+        }
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) !=
+            PackageManager.PERMISSION_GRANTED
+        ) {
+            ActivityCompat.requestPermissions(
+                this, arrayOf(Manifest.permission.ACCESS_FINE_LOCATION), PERMISSION_REQUEST_CODE
+            )
+        } else {
+            initLocationManager()
+        }
+    }
+
+    private fun initLocationManager() {
+        try {
+            locationManager = getSystemService(LOCATION_SERVICE) as LocationManager
+
+            if (!locationManager.isProviderEnabled(LocationManager.GPS_PROVIDER)) return
+
+            val location = locationManager.getLastKnownLocation(LocationManager.GPS_PROVIDER)
+            if (location != null) onLocationChanged(location)
+
+            locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, 0, 0f, this)
+        } catch (e: SecurityException) {
+            // Handle exception if needed
+        }
+    }
+
+    override fun onLocationChanged(location: Location) {
+        swipeFragment.updateLocation(location.latitude, location.longitude)
+    }
+
+    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+        if (requestCode == PERMISSION_REQUEST_CODE && grantResults.isNotEmpty() &&
+            grantResults[0] == PackageManager.PERMISSION_GRANTED
+        ) {
+            initLocationManager()
+        }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        if (::locationManager.isInitialized) {
+            locationManager.removeUpdates(this)
+        }
+    }
 }
