+++ app/src/main/java/com/example/munch_cmpt362/ui/auth/AuthViewModel.kt
@@ -1,0 +1,38 @@
+package com.example.munch_cmpt362.ui.auth
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.example.munch_cmpt362.data.repository.AuthRepository
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@HiltViewModel
+class AuthViewModel @Inject constructor(
+    private val authRepository: AuthRepository
+) : ViewModel() {
+
+    private val _loginResult = MutableLiveData<Result<Unit>>()
+    val loginResult: LiveData<Result<Unit>> = _loginResult
+
+    private val _registerResult = MutableLiveData<Result<Unit>>()
+    val registerResult: LiveData<Result<Unit>> = _registerResult
+
+    fun login(email: String, password: String) {
+        viewModelScope.launch {
+            _loginResult.value = authRepository.login(email, password)
+        }
+    }
+
+    fun register(email: String, password: String) {
+        viewModelScope.launch {
+            _registerResult.value = authRepository.register(email, password)
+        }
+    }
+
+    fun logout() {
+        authRepository.logout()
+    }
+}
